pyproject.toml
[project]
name = "agent_fleet"
version = "0.1.0"
description = "Remote development environment manager via Telegram"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bot-framework[all]==0.2.1",
    "claude-code-sdk>=0.0.25",
    "psycopg[binary]>=3.2.10",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "pyyaml>=6.0.0",
    "aio-pika>=9.5.8",
]

[dependency-groups]
dev = [
    "import-linter>=2.9",
    "mypy>=1.18.2",
    "pyright>=1.1.406",
    "pytest>=9.0.0",
    "pytest-env>=1.2.0",
    "rope>=1.14.0",
    "ruff>=0.13.3",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.pyright]
venv = ".venv"
venvPath = "."
reportUnusedVariable = false
reportUnusedImport = false

[tool.ruff]
lint.select = [
  "F", # flake8
  "E", # pycodestyle errors
  "W", # pycodestyle warnings
  "B", # flake8-bugbear
  "N", # Check your code against PEP 8 naming conventions.
  "C", # McCabe complexity.
  "T", # Print found todos.
  "UP",  # upgrade syntax for newer versions of the language.
  "ANN", # Check for missing type annotations
  "S",   # Check for security issues
  "DTZ", # Check for datetime timezone issues
]

lint.ignore = [
  "ANN204", # Missing type annotation for special methods like __init__ etc.
  "E501", # line too long
  "C901",   # function is too complex
  "ANN401", # Useless variable
  "W293",   # blank line contains whitespace
  "W291",   # trailing whitespace
]

[tool.ruff.lint.per-file-ignores]
"test*" = [
  "S101",   # disable security issues for tests
  "ANN201", # Missing type annotation for test function
]

[tool.ruff.format]
skip-magic-trailing-comma = false

[tool.mypy]
ignore_missing_imports = true

strict = false
disallow_subclassing_any = true
warn_unused_ignores = false
warn_redundant_casts = true
check_untyped_defs = true
disallow_untyped_decorators = true

disallow_any_generics = false
disallow_any_expr = false
disallow_any_decorated = false
disallow_any_explicit = false

plugins = []

show_error_codes = true

[tool.importlinter]
root_packages = ["src", "workers"]

# Глобальные слои: workers -> src
[[tool.importlinter.contracts]]
name = "Global layers"
type = "layers"
layers = ["workers", "src"]

[[tool.importlinter.contracts]]
name = "src layers"
type = "layers"
layers = ["src.flows", "src.bounded_context", "src.messaging", "src.base_protocols"]

[tool.uv.sources]
bot-framework = { path = "/Users/vladimirsumarokov/dev/smartist/bot_framework", editable = true }
